Add allowOmitUnit instead of new rule
