+++ detekt-core/src/main/resources/default-detekt-config.yml
@@ -431,9 +431,6 @@ potential-bugs:
     active: false
   ExplicitGarbageCollectionCall:
     active: true
-  ExplicitlyDefineReturn:
-    active: false
-    allowOmitUnit: true
   HasPlatformType:
     active: true
   IgnoredReturnValue:
