+++ detekt-rules-errorprone/src/main/kotlin/io/gitlab/arturbosch/detekt/rules/bugs/PotentialBugProvider.kt
@@ -26,7 +26,6 @@ class PotentialBugProvider : DefaultRuleSetProvider {
             EqualsWithHashCodeExist(config),
             ExitOutsideMain(config),
             ExplicitGarbageCollectionCall(config),
-            ExplicitlyDefineReturn(config),
             HasPlatformType(config),
             ImplicitDefaultLocale(config),
             InvalidRange(config),
